fix(tooling): plain .mjs hocon-json adapter (Copilot review)
- Convert tools/hocon-json.ts -> tools/hocon-json.mjs: plain ESM, no TypeScript
  type stripping, so it runs on the declared Node floor (engines >=22), not
  only Node >= 23.6.
- Import the built dist dynamically inside the try/catch, so a missing/invalid
  bundle yields the single-line {"__error__":..} record instead of an uncaught
  module-load stack trace.
- Use process.stdout.write + process.exitCode (not console.log + process.exit)
  so stdout drains before exit and the JSON record is not truncated when piped.
